Added spectroscopy lists to the data browser. Still a little clunky.
Because of the child like relationship of spectra with respect to the channels that they are relate to they did not fit neatly into the proxy scheme. 
An anlougous method was used. Public changes are mainly confined to new enums in GwyAppWhat and couple of functions to cataloug spectra ids.



git-svn-id: svn://svn.code.sf.net/p/gwyddion/code@7561 25697fe4-d644-49b8-9d30-be5465f2adf3